% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compgraph}
\alias{compgraph}
\title{compgraph: representing differential cell-ECM communications with alluvials}
\usage{
compgraph(result.obj = NULL, targets = FALSE, only.unique = TRUE)
}
\arguments{
\item{result.obj}{character. A list with strictly two results, each scaled, from previous calls to matricom.}

\item{targets}{logical. Whether to add target genes (and their connections) to the graph. Since it might clutter the graph, default is FALSE.}

\item{only.unique}{logical. Whether to plot only the communications (alluvials) that differ between conditions (present in one and absent in another). Default is TRUE.}
}
\value{
an alluvial plot made with ggplot2
}
\description{
compgraph: representing differential cell-ECM communications with alluvials
}
\examples{
compgraph(list(a=results_1,b=results_2))
}
