% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter.finder}
\alias{filter.finder}
\title{filter.finder: graphical utility to explore filters (expression and min.pct) for matRicom}
\usage{
filter.finder(seurat.obj = NULL, group.column = NULL)
}
\arguments{
\item{seurat.obj}{character. A Seurat, Seurat-like object, or a spatial Seurat object.}

\item{group.column}{character. Name of the column in metadata with cell identities to be used.}
}
\value{
a faceted plot showing mean gene expression and % expression per cell identity together with results from 100 random samples and the defulat thresholds.
}
\description{
filter.finder: graphical utility to explore filters (expression and min.pct) for matRicom
}
\examples{
filter.finder(seurat.obj,group.column)
}
