% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flowgraph}
\alias{flowgraph}
\title{flowgraph: representing cell-ECM communications with alluvials}
\usage{
flowgraph(
  result.obj = NULL,
  targets = TRUE,
  color.by = c("sender", "ecm", "receiver"),
  simplify = TRUE
)
}
\arguments{
\item{result.obj}{character. The results from a previous call to matricom.}

\item{targets}{logical. Whether to add target genes (and their connections) to the graph. Since it might clutter the graph, default is FALSE.}

\item{color.by}{characters. Color the ribbons (alluvials) by the sender or the receiver population, or by the ECM component. Default is "sender".}

\item{simplify}{logical. Whether to cut all triplets where a component's expression (scaled) is below 0.1. Default is TRUE, to de-clutter the graph. A failsafe mechanism is triggered if all values fall below 0.1 and this option is TRUE.}
}
\value{
an alluvial plot made with ggplot2
}
\description{
flowgraph: representing cell-ECM communications with alluvials
}
\examples{
flowgraph(results)
}
