% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes_mod.R
\name{get_expressed_genes_mod}
\alias{get_expressed_genes_mod}
\title{get_expressed_genes_mod: Get expressed genes mod}
\usage{
get_expressed_genes_mod(ident, seurat_obj, pct = 0.1, assay_oi = NULL)
}
\value{
Gets expressed genes mod
}
\description{
get_expressed_genes_mod: Get expressed genes mod
}
