% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.object}
\alias{make.object}
\title{make.object: a wrapper around the Seurat workflow to prepare 10X data for matRicom}
\usage{
make.object(
  data.name = NULL,
  data.dir = NULL,
  project.name = NULL,
  data.multiple = FALSE,
  dirty.trick = FALSE,
  min.cells = 3,
  min.features = 200,
  percent.mt = 5,
  verbose = TRUE
)
}
\arguments{
\item{data.name}{character. Internal data name.}

\item{data.dir}{character. The full path to the directory containing the objects needed for the Load10X function in Seurat}

\item{project.name}{character. How the project will be called. If NULL, data.name are used.}

\item{data.multiple}{logical. IWhether data contain multiple types. If TRUE, "Gene Expression" is used. Default is FALSE.}

\item{dirty.trick}{logical. Whether to turn gene names to uppercase in case of non-human data (matRicom would otherwise crash). Default is FALSE.}

\item{min.cells}{numeric. The minimum number of cells expressing a given feature.}

\item{min.features}{numeric. The minimum number of features per cell.}

\item{percent.mt}{numeric. Cutoff for mitochondrial RNA content.}
}
\value{
a Seurat object
}
\description{
make.object: a wrapper around the Seurat workflow to prepare 10X data for matRicom
}
\examples{
make.object(data.name,data.dir)
}
