% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.object.spatial}
\alias{make.object.spatial}
\title{make.object.spatial: a wrapper around the Seurat workflow to prepare spatial 10X data for matRicom}
\usage{
make.object.spatial(
  data.name = NULL,
  data.dir = NULL,
  project.name = NULL,
  h5.filename = NULL,
  slice.number = "slice1",
  dirty.trick = FALSE,
  percent.mt = 5,
  verbose = TRUE
)
}
\arguments{
\item{data.name}{character. Internal data name.}

\item{data.dir}{character. The full path to the directory containing the objects needed for the Load10X_spatial function in Seurat (including h5 file). The folder must contain the h5 file plus a subfolder with the other requested data.}

\item{project.name}{character. How the project will be called. If NULL, data.name are used.}

\item{h5.filename}{character. Name (only the name, without full path) of the h5 file. Must contain the file extension.}

\item{dirty.trick}{logical. Whether to turn gene names to uppercase in case of non-human data (matRicom would othewise crash)}

\item{percent.mt}{numeric. Cutoff for mitochondrial RNA content.}

\item{min.cells}{numeric. The minimum number of cells expressing a given feature.}

\item{min.features}{numeric. The minimum number of features per cell.}
}
\value{
a Seurat spatial object
}
\description{
make.object.spatial: a wrapper around the Seurat workflow to prepare spatial 10X data for matRicom
}
\examples{
make.object(data.dir,h5.filename)
}
