% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricom_ui.R
\name{matricom.app}
\alias{matricom.app}
\title{matricom.app: launch matricom as an interactive, local app}
\usage{
matricom.app()
}
\value{
a local instance of the matricom shinyapp, without connection to the (online-only) OA data
}
\description{
matricom.app: launch matricom as an interactive, local app
}
