% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netgraph}
\alias{netgraph}
\title{netgraph: representing cell-ECM communications with networks, focusing on intracellular paths}
\usage{
netgraph(
  result.obj = NULL,
  receiving.population = NULL,
  ECM.component = NULL,
  label.size = NULL,
  simplified = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{result.obj}{character. The results from a previous call to matricom.}

\item{receiving.population}{character. The name of one receiving population in the results. If NULL (default), a random population is picked.}

\item{ECM.component}{character. The name of one ECM component impacting on a receiving population in the results. If NULL (default), a random ECM component is picked.}

\item{label.size}{numeric. Size of the vertex labels in the network plot. If NULL (default), no labels are added to the nodes of the plot.}

\item{simplified}{logical. Whether to remove ECM-loops (an ECM component might also work as a receptor and/or be a target of its own stimulation). Since it helps de-cluttering the graph, default is TRUE.}

\item{interactive}{logical. Whether to use the VisNetwork package to plut an interactive form of the graph. Default is FALSE.}
}
\value{
a network plot made with igraph
}
\description{
netgraph: representing cell-ECM communications with networks, focusing on intracellular paths
}
\examples{
netgraph(results)
}
