% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes_mod.R
\name{remove_reciprocalRows}
\alias{remove_reciprocalRows}
\title{remove_reciprocalRows: Remove reciprocal rows}
\usage{
remove_reciprocalRows(data)
}
\value{
Removes reciprocal rows
}
\description{
remove_reciprocalRows: Remove reciprocal rows
}
