% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatplot}
\alias{spatplot}
\title{Spatplot: representing cell-ECM communication in spatial RNAseq data}
\usage{
spatplot(
  result.obj = NULL,
  seurat.obj = NULL,
  group.column = NULL,
  feature = NULL,
  singles = FALSE
)
}
\arguments{
\item{result.obj}{character. The results from a previous call to matricom.spatial.}

\item{seurat.obj}{character. A spatial Seurat object.}

\item{group.column}{character. Name of the column in metadata with cell identities to be used.}

\item{feature}{character. A quadruplet to be plotted, collapsed by underscores (e.g.,"gene_receptor_target_TF"). If NULL (default), the function will pick a random quadruplet.}

\item{singles}{logical. Whether to plot also spatial plots of the ECM/receptor/target genes. Default is FALSE}
}
\value{
a multiplot with at least a spatial cluster plot, a spatial plot with clusters showing the chosen signal, and a spatial plot with average ECM/receptor/target intensities per dot. Might also include the spatial plots of the single elements.
}
\description{
Spatplot: representing cell-ECM communication in spatial RNAseq data
}
\examples{
spatplot(results,obj,"group.column")
}
